+++ quiz/execution.py
@@ -15,8 +15,11 @@ __all__ = [
     'execute_async',
     'executor',
     'async_executor',
-    'ErrorResponse',
+
     'Executable',
+
+    'ErrorResponse',
+    'HTTPError',
 ]
 
 Executable = t.Union[str, Query]
@@ -41,10 +44,15 @@ def _exec(executable):
 @py2_compatible
 def middleware(url, query_str):
     # type: (str, str) -> snug.Query[Dict[str, JSON]]
-    response = yield snug.Request('POST', url, json.dumps({
-        'query': query_str,
-    }).encode('ascii'), headers={'Content-Type': 'application/json'})
-    content = json.loads(response.content.decode())
+    response = yield snug.Request(
+        'POST',
+        url,
+        content=json.dumps({'query': query_str}).encode('ascii'),
+        headers={'Content-Type': 'application/json'}
+    )
+    if response.status_code >= 400:
+        raise HTTPError(response)
+    content = json.loads(response.content.decode('utf-8'))
     if 'errors' in content:
         raise ErrorResponse(**content)
     return_(content['data'])
@@ -72,6 +80,8 @@ def execute(obj, url, **kwargs):
     ------
     ErrorResponse
         If errors are present in the response
+    HTTPError
+        If the response has a non 2xx response code
     """
     snug_query = irelay(_exec(obj), partial(middleware, url))
     return snug.execute(snug_query, **kwargs)
@@ -130,6 +140,8 @@ def execute_async(obj, url, **kwargs):
     ------
     ErrorResponse
         If errors are present in the response
+    HTTPError
+        If the response has a non 2xx response code
     """
     snug_query = irelay(_exec(obj), partial(middleware, url))
     return snug.execute_async(snug_query, **kwargs)
@@ -171,3 +183,13 @@ class ErrorResponse(ValueObject, Exception):
         ('errors', t.List[t.Dict[str, JSON]],
          'Errors returned in the response'),
     ]
+
+
+class HTTPError(ValueObject, Exception):
+    __fields__ = [
+        ('response', snug.Response, 'The response object'),
+    ]
+
+    def __str__(self):
+        return ('Response with status {0.status_code}, content: {0.content!r}'
+                .format(self.response))