+++ tests/test_execution.py
@@ -83,6 +83,14 @@ class TestExecute:
         assert exc.value == quiz.ErrorResponse({'foo': 4},
                                                [{'message': 'foo'}])
 
+    def test_http_error(self):
+        err_response = snug.Response(403, b'this is an error!')
+        client = MockClient(err_response)
+        with pytest.raises(quiz.HTTPError) as exc:
+            quiz.execute('my query', url='https://my.url/api',
+                         client=client, auth=token_auth('foo'))
+        assert exc.value == quiz.HTTPError(err_response)
+
 
 def test_executor():
     executor = quiz.executor(url='https://my.url/graphql')
@@ -162,3 +170,9 @@ def test_async_executor():
     executor = quiz.async_executor(url='https://my.url/graphql')
     assert executor.func is quiz.execute_async
     assert executor.keywords['url'] == 'https://my.url/graphql'
+
+
+def test_http_error():
+    err = quiz.HTTPError(snug.Response(404, content=b'not found!\x00'))
+    assert 'not found!\\x00' in str(err)
+    assert '404' in str(err)